Fixed a failing test
Test was failing because of a deps update: @types/node
The test was depending @types/node to be at version 15.0.1
but at the time of this commit @types/node was 15.6.0.

I have added a dev dependency (@types/mssql@6.0.8) and set
the package.json deps entry to match this version exactly.

Signed-off-by: Mehmet Baker <mehmet@mehmetbaker.dev>