bug: Fix bug in excludePackagesStartingWith function
The it worked before, if you added more than one name to the blacklist,
then while it was checking for the first name in the list, the
subsequent names would still be put into the resultJson object.

Now, I first check if the current package is one of all blacklisted
names and then, depending on the result, either add it to resultJson or
remove it from the filtered object.