config: Add NodeJS and npm version information to package.json
Unfortunately, this detail does not really what I want it to, but maybe
it's also a gift for now, as it does not prevent older NodeJS and npm
versions to work with release 4.0.0, which will of course work.

But still, since package.json offers this kind of information, I make
use of it.

Also, I re-ordered the attributes of the JSON file, as I felt that the
dependencies should go last. Might be that I (or you) will change my
mind on this again.