fix: Fix choosing the wrong license info
A "custom" license should be accepted, I guess. If this line is not
commented out, it leads to "custom" licenses from the package.json
to be overridden by any URL written in the first line of the README, if
applicable. And that is definitely a bug. So this commented out line
leads to an improvement for now.
Nevertheless, the code should be improved, so that e.g. README contents
are being parsed in a better way and then the code decides which license
info to use.