refactor: Rename regular expression constants and add one check
Improve the detection of URLs as licenses which are no licenses at all.
Previously, when no license info could be found elsewhere, any URL in
the README was taken as a custom license, which is not a very
bulletproof method. Now, I restrict this method which was probably meant
as a fallback solution to only being considered if the README contains
at least the word "license" in some form (or notation). Not good, but
better than before.