Add files allow-list in package.json
The current released version of the package contains a "test"
directory with a total size of 146.9MiB. It contains a bunch of
unrelated NPM packages in a node_modules directory, for instance
angular.

I suspect this directory is local to the maintainer's machine and
not meant to be included in NPM releases. To prevent this from
happening, add a "files" allow-list to only include what's needed
for the package to run in the output of "npm pack".